ENH: Calculator Solve Functionality Added
- "Solve" functionality added with the `=` button.
- Calculator can now solve basic algebraic field operations in any base
system from `0` to `16`.
- During testing a bug was discovered that the root selection enabling
of the buttons is actually uninitialized and the root selection itself
is enabling an extra button (off-by-1 indexing error).
- Simple examples tested and functionality seems to be working for the
`+`, `-`, and `*` operations of integers. Fractional values and division
have not yet been tested/verified.
- Current design is likely incompatible with fractional values, so it
may be more prudent to remove the decimal button from the UI for now.